+++ zebra-state/src/lib.rs
@@ -37,7 +37,8 @@ mod tests;
 
 pub use config::{
     check_and_delete_old_databases, check_and_delete_old_state_databases,
-    database_format_version_on_disk, state_database_format_version_on_disk, Config,
+    database_format_version_on_disk, state_database_format_version_on_disk, Config, PruningConfig,
+    StorageMode,
 };
 pub use constants::{state_database_format_version_in_code, MAX_BLOCK_REORG_HEIGHT};
 pub use error::{